refactor(visor): split hypervisor.go HTTP handlers by topic
hypervisor.go was 2385 lines mostly because of 78 HTTP handler methods
on *Hypervisor covering every endpoint the GUI/CLI calls. Lifecycle and
core wiring (Enable/Disable/IsEnabled, ServeRPC over DMSG, makeMux chi
router, HTTPHandler) actually only account for ~700 lines — the rest is
flat handler functions grouped only by /* SECTION */ comments.

Reassessed the option of moving to pkg/hypervisor: hypervisor and visor
are tightly coupled by design (no standalone hypervisor anymore;
listener and config are persistent on the visor; hypervisor TUI also
runs in-binary). Extracting to a separate package would force exporting
internal Visor fields with no isolation gain. Instead, do an in-package
split mirroring rpc.go and redis_store.go.

Topic files (all package visor):
  hypervisor.go                            (698)  struct, Enable, Disable,
                                                  IsEnabled, ServeRPC,
                                                  makeMux, HTTPHandler,
                                                  log, types, AddMockData
  hypervisor_handlers_admin.go             (232)  rewards, log rotation,
                                                  isPublic, runtime config,
                                                  shutdown, runtime logs
  hypervisor_handlers_apps.go              (236)  apps list/detail/stats,
                                                  putApp, logs, connections
  hypervisor_handlers_ctx.go               (165)  visorConn, withCtx,
                                                  visorCtx, appCtx, tpCtx,
                                                  routeCtx
  hypervisor_handlers_dmsg.go              (119)  dmsg + LAN dmsg server +
                                                  sessions count
  hypervisor_handlers_health.go            (75)   getHealth/Service/Uptime
  hypervisor_handlers_misc.go              (62)   ping, csrf, about, pty
  hypervisor_handlers_proxies.go           (63)   embedded proxies status,
                                                  enable/upstream
  hypervisor_handlers_routes.go            (215)  routes, route groups,
                                                  min hops
  hypervisor_handlers_skynet.go            (184)  skynet ports, forwarded
                                                  ports, skynet forwards
  hypervisor_handlers_transports.go        (197)  transports CRUD,
                                                  persistent transports,
                                                  public_autoconnect
  hypervisor_handlers_visors.go            (209)  getVisors,
                                                  getVisor[Summary],
                                                  getAllVisorsSummary

No behavior change.
