+++ pkg/visor/hypervisor_handlers_dmsg.go
@@ -1,0 +1,119 @@
+// Package visor pkg/visor/hypervisor.go
+package visor
+
+import (
+	"context"
+	"net/http"
+	"time"
+
+	"github.com/skycoin/skywire/pkg/skywire-utilities/pkg/cipher"
+	"github.com/skycoin/skywire/pkg/skywire-utilities/pkg/httputil"
+	"github.com/skycoin/skywire/pkg/visor/dmsgtracker"
+)
+
+func (hv *Hypervisor) getLANDmsgServer() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		if hv.lanDmsg == nil {
+			httputil.WriteJSON(w, r, http.StatusOK, LANDmsgServerInfo{Enabled: false})
+			return
+		}
+		httputil.WriteJSON(w, r, http.StatusOK, LANDmsgServerInfo{
+			Enabled: true,
+			PK:      hv.lanDmsg.PK,
+			Address: hv.lanDmsg.Address,
+		})
+	}
+}
+
+func (hv *Hypervisor) getDmsg() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		out := hv.getDmsgSummary()
+		httputil.WriteJSON(w, r, http.StatusOK, out)
+	}
+}
+
+// getDmsgSessions returns per-client dmsg session info: main + embedded
+// route setup node + embedded transport setup node. Used by the hypervisor
+// UI's dmsg settings panel so operators can see at a glance which dmsg
+// servers each of the visor's three independent dmsg clients is connected
+// to.
+func (hv *Hypervisor) getDmsgSessions() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		if hv.visor == nil {
+			httputil.WriteJSON(w, r, http.StatusServiceUnavailable, &DmsgClientSessions{})
+			return
+		}
+		sessions, err := hv.visor.DmsgSessions()
+		if err != nil {
+			httputil.WriteJSON(w, r, http.StatusInternalServerError, map[string]string{"error": err.Error()})
+			return
+		}
+		httputil.WriteJSON(w, r, http.StatusOK, sessions)
+	}
+}
+
+// postDmsgConnectAll triggers a one-shot "reach every dmsg server now"
+// action on the main visor dmsg client. Used by the UI's "Connect to all"
+// button for operators running RSN/TPS visors who want to eliminate
+// discovery-stale failure modes on demand.
+func (hv *Hypervisor) postDmsgConnectAll() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		if hv.visor == nil {
+			httputil.WriteJSON(w, r, http.StatusServiceUnavailable, map[string]string{"error": "visor not available"})
+			return
+		}
+		result, err := hv.visor.DmsgConnectAll()
+		if err != nil {
+			httputil.WriteJSON(w, r, http.StatusInternalServerError, map[string]string{"error": err.Error()})
+			return
+		}
+		httputil.WriteJSON(w, r, http.StatusOK, result)
+	}
+}
+
+// putDmsgSessionsCount persists dmsg.sessions_count to the visor config and
+// triggers an immediate connect-all so the change takes effect without a
+// restart. A value of 0 means "connect to every available dmsg server",
+// which is the recommended setting for RSN/TPS visors.
+func (hv *Hypervisor) putDmsgSessionsCount() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		if hv.visor == nil {
+			httputil.WriteJSON(w, r, http.StatusServiceUnavailable, map[string]string{"error": "visor not available"})
+			return
+		}
+		var req dmsgSessionsCountRequest
+		if err := httputil.ReadJSON(r, &req); err != nil {
+			httputil.WriteJSON(w, r, http.StatusBadRequest, map[string]string{"error": err.Error()})
+			return
+		}
+		if req.Count < 0 {
+			httputil.WriteJSON(w, r, http.StatusBadRequest, map[string]string{"error": "count must be >= 0"})
+			return
+		}
+		result, err := hv.visor.SetDmsgSessionsCount(req.Count)
+		if err != nil {
+			httputil.WriteJSON(w, r, http.StatusInternalServerError, map[string]string{"error": err.Error()})
+			return
+		}
+		httputil.WriteJSON(w, r, http.StatusOK, result)
+	}
+}
+
+func (hv *Hypervisor) getDmsgSummary() []dmsgtracker.DmsgClientSummary {
+	hv.mu.RLock()
+	pks := make([]cipher.PubKey, 0, len(hv.remoteVisors)+1)
+	if hv.visor != nil {
+		pks = append(pks, hv.visor.conf.PK)
+	}
+	for pk := range hv.remoteVisors {
+		pks = append(pks, pk)
+	}
+	hv.mu.RUnlock()
+
+	if hv.visor.isDTMReady() {
+		ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+		defer cancel()
+		return hv.visor.dmsgTracker.manager.GetBulk(ctx, pks)
+	}
+	return []dmsgtracker.DmsgClientSummary{}
+}
