+++ pkg/visor/hypervisor_handlers_health.go
@@ -1,0 +1,75 @@
+// Package visor pkg/visor/hypervisor.go
+package visor
+
+import (
+	"net/http"
+	"time"
+
+	"github.com/skycoin/skywire/pkg/skywire-utilities/pkg/httputil"
+)
+
+// getServiceHealth returns the health status of all configured deployment
+// services (TPD, DMSG discovery, AR, RF, UT, SD). Each entry has name, URL,
+// status, latency and version. Fetched via the local visor's DMSG/HTTP
+// client so the UI sees the same view the visor does.
+func (hv *Hypervisor) getServiceHealth() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		if hv.visor == nil {
+			httputil.WriteJSON(w, r, http.StatusServiceUnavailable, []ServiceHealthEntry{})
+			return
+		}
+		entries, err := hv.visor.ServiceHealth()
+		if err != nil {
+			httputil.WriteJSON(w, r, http.StatusInternalServerError, map[string]string{"error": err.Error()})
+			return
+		}
+		httputil.WriteJSON(w, r, http.StatusOK, entries)
+	}
+}
+
+// provides summary of health information for every visor
+func (hv *Hypervisor) getHealth() http.HandlerFunc {
+	return hv.withCtx(hv.visorCtx, func(w http.ResponseWriter, r *http.Request, ctx *httpCtx) {
+		vh := &Health{}
+
+		type healthRes struct {
+			h   *HealthInfo
+			err error
+		}
+
+		resCh := make(chan healthRes)
+		tCh := time.After(HealthTimeout)
+
+		go func() {
+			hi, err := ctx.API.Health()
+			resCh <- healthRes{hi, err}
+		}()
+
+		select {
+		case res := <-resCh:
+			if res.err != nil {
+				vh.Status = http.StatusInternalServerError
+			} else {
+				vh.HealthInfo = res.h
+				vh.Status = http.StatusOK
+			}
+
+			httputil.WriteJSON(w, r, http.StatusOK, vh)
+		case <-tCh:
+			httputil.WriteJSON(w, r, http.StatusRequestTimeout, &Health{Status: http.StatusRequestTimeout})
+		}
+	})
+}
+
+// getUptime gets given visor's uptime
+func (hv *Hypervisor) getUptime() http.HandlerFunc {
+	return hv.withCtx(hv.visorCtx, func(w http.ResponseWriter, r *http.Request, ctx *httpCtx) {
+		u, err := ctx.API.Uptime()
+		if err != nil {
+			httputil.WriteJSON(w, r, http.StatusInternalServerError, err)
+			return
+		}
+
+		httputil.WriteJSON(w, r, http.StatusOK, u)
+	})
+}
